server/pyproject.toml
[tool.poetry]
name = "test-server"
version = "0.1.0"
description = ""
authors = ["Marce Coll <marce@dziban.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg = "^3.1.8"
fastapi = "^0.91.0"
uvicorn = "^0.20.0"
pydantic = "^1.10.4"
psycopg-pool = "^3.1.6"
jinja2 = "^3.1.2"
fastapi-login = "^1.8.3"
python-multipart = "^0.0.5"

[tool.poetry.group.dev.dependencies]
ipython = "^8.10.0"
pytest = "^7.2.2"
pytest-playwright = "^0.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
